# src/main/java/com/example/demo/Model/Subject.java
package com.example.demo.Model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Subject {

	private Integer subId;
	private String subName;
	private String teacherName;
}